Created AssignedExams for the many-to-many relationship between Exams and Users
I chose to use has_many :through in this case. This isn't strictly necessary for the requirements of this exercise, but seeing how this model would actually be used, I fully expect that we would start tracking additional metadata on the relationship. For example, we could track if the user showed up for the exam.